requirements.txt
python-dotenv==1.1.1
requests==2.32.5
youtube-transcript-api==1.2.2
yt-dlp==2025.9.26 
openai-whisper==20250625